requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.12.1

# Reddit
praw==7.7.1

# ML/AI (without LightGBM for now)
torch==2.1.1
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
# lightgbm==4.1.0  # Commented out - we'll use XGBoost or sklearn's GradientBoosting instead
xgboost==2.0.3  # Alternative to LightGBM
faiss-cpu==1.7.4

# Data Processing
pandas==2.1.3
scipy==1.11.4

# Utils
tqdm==4.66.1

# Monitoring
prometheus-client==0.19.0